requirements.txt
alembic==1.7.7
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.1.1
Flask-HTTPAuth==4.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pause==0.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
six==1.16.0
SQLAlchemy==1.4.35
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
